Update file with conditon

+++ shell/misc.sh
@@ -2,5 +2,8 @@
 
 echo -e "Updating miscellaneous configuration files..."
 
+
+if [[ "$OSTYPE" != 'linux-gnu' ]]; then
 rm -f ~/.hushlogin
 ln -s ~/dotfiles/.hushlogin ~/.hushlogin
+fi
